hw7-test

+++ src/services/auth.js
@@ -1,12 +1,6 @@
 import createHttpError from 'http-errors';
-import bcrypt from 'bcrypt';
-import jwt from 'jsonwebtoken';
-import path from 'node:path';
-import fs from 'node:fs/promises';
-import handlebars from 'handlebars';
-import { randomBytes } from 'crypto';
-
 import { UsersCollection } from '../db/models/user.js';
+import bcrypt from 'bcrypt';
 import { SessionsCollection } from '../db/models/session.js';
 import {
   FIFTEEN_MINUTES,
@@ -14,7 +8,12 @@ import {
   TEMPLATES_DIR,
   THIRTY_DAYS,
 } from '../constants/index.js';
+import { randomBytes } from 'crypto';
+import jwt from 'jsonwebtoken';
 import { getEnvVar } from '../utils/getEnvVar.js';
+import path from 'node:path';
+import fs from 'node:fs/promises';
+import handlebars from 'handlebars';
 import { sendResetEmail } from '../utils/sendResetEmail.js';
 
 const createSession = () => {
@@ -84,6 +83,8 @@ export const logoutUser = async (refreshToken) => {
   const session = await SessionsCollection.findOne({
     refreshToken,
   });
+  if (!session) throw createHttpError(401, 'session not found');
+
   await SessionsCollection.deleteOne({ _id: session._id, refreshToken });
 };
